+++ crates/block-producer/src/lib.rs
@@ -10,7 +10,7 @@ mod block_builder;
 mod domain;
 mod mempool;
 pub mod store;
-pub mod validator;
+mod validator;
 
 #[cfg(feature = "testing")]
 pub mod errors;
@@ -49,6 +49,24 @@ const SERVER_MEMPOOL_EXPIRATION_SLACK: u32 = 2;
 
 /// The interval at which to update the cached mempool statistics.
 const CACHED_MEMPOOL_STATS_UPDATE_INTERVAL: Duration = Duration::from_secs(5);
+
+/// How often a block is created.
+pub const DEFAULT_BLOCK_INTERVAL: Duration = Duration::from_secs(5);
+
+/// How often a batch is created.
+pub const DEFAULT_BATCH_INTERVAL: Duration = Duration::from_secs(2);
+
+/// The default transaction capacity of the mempool.
+///
+/// The value is selected such that all transactions should approximately be processed within one
+/// minutes with a block time of 5s.
+#[allow(clippy::cast_sign_loss, reason = "Both durations are positive")]
+pub const DEFAULT_MEMPOOL_TX_CAPACITY: NonZeroUsize = NonZeroUsize::new(
+    DEFAULT_MAX_BATCHES_PER_BLOCK
+        * DEFAULT_MAX_TXS_PER_BATCH
+        * (Duration::from_secs(60).div_duration_f32(DEFAULT_BLOCK_INTERVAL)) as usize,
+)
+.unwrap();
 
 const _: () = assert!(
     DEFAULT_MAX_BATCHES_PER_BLOCK <= miden_objects::MAX_BATCHES_PER_BLOCK,